refactor: :recycle: refactoriza logica de validacion en un solo archivo utils
